[REF] web_tour: inline getScrollParent and isInPage in tour_pointer
tour_pointer.js was the only remaining consumer of tour_utils.js.
