+++ addons/web_tour/static/src/js/tour_pointer/tour_pointer.js
@@ -5,7 +5,6 @@ import { browser } from "@web/core/browser/browser";
 import { usePosition } from "@web/core/position/position_hook";
 import { _t } from "@web/core/l10n/translation";
 import { usePopover } from "@web/core/popover/popover_hook";
-import { getScrollParent, isInPage } from "@web_tour/js/utils/tour_utils";
 
 const oppositeSides = {
     left: "right",
@@ -13,6 +12,47 @@ const oppositeSides = {
     top: "bottom",
     bottom: "top",
 };
+
+/**
+ * @param {HTMLElement} element
+ * @returns {HTMLElement | null}
+ */
+function getScrollParent(element) {
+    if (!element) {
+        return null;
+    }
+    // We cannot only rely on the fact that the element’s scrollHeight is
+    // greater than its clientHeight. This might not be the case when a step
+    // starts, and the scrollbar could appear later. For example, when clicking
+    // on a "building block" in the "building block previews modal" during a
+    // tour (in website edit mode). When the modal opens, not all "building
+    // blocks" are loaded yet, and the scrollbar is not present initially.
+    const overflowY = window.getComputedStyle(element).overflowY;
+    const isScrollable =
+        overflowY === "auto" ||
+        overflowY === "scroll" ||
+        (overflowY === "visible" && element === element.ownerDocument.scrollingElement);
+    if (isScrollable) {
+        return element;
+    } else {
+        return getScrollParent(element.parentNode);
+    }
+}
+
+function isInPage(element) {
+    if (!element || !element.isConnected) {
+        return false;
+    }
+    const doc = element.ownerDocument;
+    if (doc === document) {
+        return document.body.contains(element);
+    }
+    if (doc.defaultView && doc.defaultView.frameElement) {
+        const iframe = doc.defaultView.frameElement;
+        return document.body.contains(iframe);
+    }
+    return false;
+}
 
 /**
  * @typedef TourPointerState
